Add option `only_status` (#210)
Developers sometimes leave the workflows in queued state. This happens
when you have a protected environment and deployment is waiting for
approval. They sometimes do not approve and just leave the workflow
pending waiting for a merge that will start a new workflow run and
deploy from there multiple merges at once.

We need to cancel the older pending ones if new PR is merged but SKIP
canceling workflows that are running as it might be dangerous to
interrupt it. To play it safe we introduce `only_status: 'waiting'` in
order to avoid canceling runs with status `in_progress`

GitHub official implementation also does not support this option and
they do not have plans to implement it soon, so we are adding it to this
action

---------

Co-authored-by: 8666 <vlatko.boshkoski@slicelife.com>
Co-authored-by: Steven <steven@ceriously.com>